+++ CHANGELOG.md
@@ -26,7 +26,7 @@
 - Fixed tooltip overflow on progress bars during hover and scroll of uploaded items.
 - Fixed accordion to apply square prop
 - Fixed ref forwarding support for the paper component
-- Fixed the accessibility of help icon and added a new prop to enable a hover effect on it
+- Added a new prop - enableHelpHoverEffect - to improve accessibility of help icon in input field components (Textfield, Select Single/Multiple, Autocomplete, MultipleSelectChip, DatePicker).
 
 ### Changed
 - Upgraded dependency to Enchanted Icons v1.5.0